Autocall `install()` if `COLOREDLOGS_AUTOUSE=1`.
Installs a .pth file so that one can automatically call `install()` by
setting the `COLOREDLOGS_AUTOUSE` environment variable.

Note that due to the order in which `pth` files are loaded, this will not work
if `humanfriendly` is installed system-wide but `colorlogs` installed as a
user package (this is what the `except ImportError` clause catches); however
everything works fine if they are installed to the same site-packages.