Backwards incompatible: Change log level handling
The old situation:

 - install() would unconditionally change the log level of the root
   logger to logging.NOTSET (changing it from the root logger's
   default level logging.WARNING) and the log levels of handler(s)
   would control which log messages were actually emitted.

 - enable_system_logging() would never change the root logger's
   log level, which meant that when enable_system_logging() was
   used in isolation from install() the default log level would
   implicitly be set to logging.WARNING (the default log level
   of the root logger).

Over the years I've gotten a lot of feedback about the log level
handling in the coloredlogs package, it was clearly the number
one cause of confusion for users. Here are some examples:

- https://github.com/xolox/python-coloredlogs/issues/14
- https://github.com/xolox/python-coloredlogs/issues/18
- https://github.com/xolox/python-coloredlogs/pull/21
- https://github.com/xolox/python-coloredlogs/pull/23
- https://github.com/xolox/python-coloredlogs/issues/24

My hope is that with the changes I've made in the past days, the
experience for new users will be more `Do What I Mean' and less
`What The Fuck is Going On Here?!' :-). Of course only time (and
feedback, or lack thereof) will tell whether I've succeeded.