Continued: Testing system logging on MacOS (attempt 4)
In commit [1] I blamed the failing system logging tests on MacOS on
timing issues. My workaround turned out not to fix the issue at hand.

Assuming that timing issues were _not_ to blame I removed the workaround
and shifted my focus to the log levels used in the test suite [2]. This
seems to have had a positive effect [3] in the sense that two of the
three tests are now passing instead of one.

However that doesn't explain why the third test is still failing and the
only reason I can think of is … timing issues :-). So now I'm going to
combine both of my previous commits in a final attempt to reach >= 90%
test coverage on MacOS via Travis CI.

[1] 7350ea18379676ff852d30954fa976750783d69b
[2] 477d4184f1908a7ad0a071d9f0e619994673769b
[3] https://travis-ci.org/xolox/python-coloredlogs/jobs/325279489