Release 8.0: Set default date format in formatTime()
I decided to bump the major version number because this change is
technically backwards incompatible, even though it concerns a minor
detail (datefmt being set to its default value in the initializer).

Then again, this change does improve compatibility with the behavior
of the logging.Formatter class in Python's standard library,
so there's that :-).