Document now fully optional Colorama dependency (#71, #72)
As far as I can tell wheel distributions make it impossible to implement
a conditional dependency on Colorama based on the Windows OS version, so
in the spirit of looking forward (Windows 10 has a far bigger market
share than it's predecessors) I've decided to drop the requirement.

Hopefully most existing users will not experience any regression because
they previously installed Colorama already, and when available it will
still be used. However new installations on older Windows systems now
require users to install Colorama separately. This has been documented
in the readme in an attempt to minimize resulting confusion.