if user defines use_colors, then it does not try to detect terminal capabilities, it just uses it.
If the user defines use_colors, then it does not try to detect terminal capabilities, it just uses it.

This solves some scenarios like logging to consoles that do not report correctly if they support ANSI colors or not. Example: Jenkins, IntelliJ.
If the user wants the to use the autodetect capabilities, then use_colors should be None.