Bug fix: Restore StandardErrorHandler functionality
The StandardErrorHandler class is responsible for dynamically resolving
(looking up the value of) sys.stderr for each logged message instead of
once when coloredlogs.install() is called.

This was unintentionally broken by changes in 4f8c9aa8ad4 which were
released as part of coloredlogs version 14.1.

This was brought to my attention by @hugovk here:
https://github.com/xolox/python-humanfriendly/pull/51